Break circular dependency
